+++ hstratus-auth/ChangeLog.md
@@ -2,6 +2,10 @@
 
 `icloud-auth` uses [PVP Versioning][1].
 
+## 0.2.0.0 -- 2026-07-30
+
+* Remove `hstratus-auth-cli` sublibrary; move `Network.HStratus.Http.Cli` into the main library.
+
 ## 0.1.0.1 -- 2026-07-28
 
 * Update package metadata.