+++ hstratus/ChangeLog.md
@@ -2,6 +2,11 @@
 
 `hstratus` uses [PVP Versioning][1].
 
+## 0.1.1.0 -- 2026-07-30
+
+* Add `Hstratus.Cli.Common` (parser for shared CLI options).
+* Remove dependency on `hstratus-auth-cli` sublibrary.
+
 ## 0.1.0.0 -- 2026-07-28
 
 * Initial version.